PeaceFounder.Server.Controllers

Registrar

PeaceFounder.Server.Controllers.Registrar — Type

struct Registrar
    metadata::Ref{Vector{UInt8}} 
    tickets::Vector{Ticket}
    signer::Signer
    hmac::HMAC
end

Represents a state for token registrar service. To initialize the service it's necessary to create a signer who can issue a valid admisssion certificates and a secret key with which a recruit client can exchange authorized messages. See also method generate(Registrar, spec).

Metadata is used as means to securelly deliver to the client most recent server specification.

Interface: select, hmac, hasher, key, id, tickets, in, enlist!, admit!, isadmitted, ticket_status

PeaceFounder.Server.Controllers.Ticket — Type

mutable struct Ticket
    const ticketid::TicketID
    timestamp::DateTime
    attempt::UInt8
    token::Digest
    tokenid::String # 
    admission::Union{Admission, Nothing}
end

Represents a ticket state for ticket with ticketid. timestamp represents time when the ticket have been issued by a registrar client in authorization system of choice, for instance, Registrars.jl; attempt is a counter with which token can be reset which is calculated with token(ticketid, attempt, hmac). Lastly admission contains a certified member pseudonym which was authetificated by the user with token.

Base.in — Method

in(ticketid::TicketID, registrar::Registrar)::Bool

Return true if there already is a ticket with ticketid.

PeaceFounder.Core.Model.generate — Method

generate(Registrar, spec::CryptoSpec)

Generate a new token registrar with unique signer and athorization key.

PeaceFounder.Core.Model.select — Method

select(Admission, ticketid::TicketID, registrar::Registrar)::Union{Admission, Nothing}

Return admission for a ticket with given a given identity pseudonym from registrar. If no ticket with given id is found OR ticket is not yet admitted returns nothing.

PeaceFounder.Core.Model.select — Method

select(Admission, ticketid::TicketID, registrar::Registrar)::Union{Admission, Nothing}

Return admission for a ticket with given ticketid from registrar. If no ticket with given ticketid is found OR ticket is not yet admitted returns nothing.

PeaceFounder.Core.Model.select — Method

select(T, predicate::Function, registrar::Registrar)::Union{T, Nothing}

From a list of all registrar tickets return T <: Union{Ticket, Admission} for which predicate is true. If none succeds returns nothing.

PeaceFounder.Core.ProtocolSchema.isadmitted — Method

isadmitted(ticketid::TicketID, registrar::Registrar)

Check whether a ticket is already admitted. Returns false when either ticket is nonexistent or it's admission is nothing.

PeaceFounder.Server.Controllers.admit! — Method

admit!(registrar::Registrar, id::Pseudonym, ticketid::TicketID)::Admission

Attempt to admit an identity pseudonym id for ticket ticketid. Authorization is expected to happen at the service layer using provided token in the invite. If a ticket is already registered return admission if it matches the provided id. Otherwise throe an error.

PeaceFounder.Server.Controllers.enlist! — Method

enlist!(registrar::Registrar, ticketid::TicketID, timestamp::DateTime; route::URI=registrar.route)::Invite

Registers a new ticket with given TicketID and returns an invite. If ticket is already registered the same invite is returned. Throws an error when ticket is already registered.

PeaceFounder.Server.Controllers.hmac — Method

hmac(x)::HMAC

Return HMAC authorizer from a given object.

PeaceFounder.Server.Controllers.set_demehash! — Method

set_demespec!(registrar::Registrar, spec::Union{Digest, DemeSpec})

Replace metadata for a registrar. Note when data is replaced all unfinalized tokens need to be flushed.

PeaceFounder.Server.Controllers.ticket_status — Method

ticket_status(ticketid::TicketID, registrar::Registrar)::Union{TicketStatus, Nothing}

Return a ticket status for a ticketid. In case ticket is not found return nothing.

PeaceFounder.Server.Controllers.tickets — Method

tickets(registrar::Registrar)::Vector{TicketID}

Return a list of registered ticket ids.

PeaceFounder.Server.Controllers.token — Method

token(ticketid::TicketID, hmac::HMAC)

Compute a recruit token for a given ticketid. Calculates it as token=Hash(Hash(0|key)|attempt|ticketid) where attempt is a counter for which token is issued.

Note: the token generation from key is made in order to support it's local computation on a remote server where QR code for registration is shown within organization website.

BraidChain

PeaceFounder.Server.Controllers.BraidChainController — Type

struct BraidChain
    members::Set{Pseudonym}
    ledger::BraidChainLedger
    spec::DemeSpec
    generator::Generator
    tree::HistoryTree
    commit::Union{Commit{ChainState}, Nothing}
end

Represents a braidchain ledger with it's associated state. Can be instantitated with a demespec file using BraidChain(::DemeSpec) method.

Interface: push!, record!, state, length, list, select, roll, constituents, generator, commit, commit_index, ledger, leaf, root, ack_leaf, ack_root, members, commit!

Base.push! — Method

push!(ledger::BraidChainController, t::Transaction)

Add an element to the BraidChainController bypassing transaction verification with the chain. This should only be used when the ledger is loaded from a trusted source like a local disk or when final root hash is validated with a trusted source.

HistoryTrees.leaf — Method

leaf(ledger::BraidChainController, N::Int)::Digest

Return a ledger's element digest at given index.

HistoryTrees.root — Method

root(ledger::BraidChainController[, N::Int])::Digest

Return a ledger root digest. In case when index is not given a current index is used.

PeaceFounder.Core.Model.commit — Method

commit(ledger::BraidChainController)

Return a current commit for a braichain.

PeaceFounder.Core.Model.generator — Method

generator(ledger[, index])

Return a generator at braidchain ledger row index. If index is omitted return the current state value.

PeaceFounder.Core.Model.generator — Method

generator(ledger::BraidChainController)

Return a current relative generator for a braidchain ledger.

PeaceFounder.Core.Model.isbinding — Method

isbinding(chain::BraidChainController, state::ChainState)::Bool

Check that chain state is consistent with braidchain ledger.

PeaceFounder.Core.Model.members — Method

members(chain::BraidChainController, [n::Int])::Set

Return a set of member pseudonyms which at given anchor index can participate in voting or braiding.

PeaceFounder.Core.Model.select — Method

select(T, predicate::Function, ledger::BraidChainController)::Union{T, Nothing}

Return a first element from a ledger with a type T which satisfies a predicate.

PeaceFounder.Core.Model.state — Method

state(ledger::BraidChainController)

Return a current braidchain ledger state metadata.

PeaceFounder.Server.Controllers.ack_leaf — Method

ack_leaf(ledger::BraidChainController, index::Int)::AckInclusion

Return a proof for record inclusion with respect to a current braidchain ledger history tree root.

PeaceFounder.Server.Controllers.ack_root — Method

ack_root(ledger::BraidChainController, index::Int)

Return a proof for the ledger root at given index with respect to the current braidchain ledger history tree root.

PeaceFounder.Server.Controllers.commit! — Method

commit!(ledger::BraidChainController, signer::Signer)

Commit a current braidchain ledger state with a signer's issued cryptographic signature.

PeaceFounder.Server.Controllers.constituents — Method

constituents(ledger::BraidChainController)::Set{Pseudonym}

Return all member identity pseudonyms.

PeaceFounder.Server.Controllers.list — Method

list(T, ledger::BraidChainController)::Vector{Tuple{Int, T}}

List braidchain elements with a given type together with their index.

PeaceFounder.Server.Controllers.reset_tree! — Method

reset_tree!(ledger::BraidChainController)

Recompute a chain tree hash.

PeaceFounder.Server.Controllers.roll — Method

roll(ledger::BraidChainController)::Vector{Membership}

Return all member certificates from a braidchain ledger.

BallotBox and PollingStation

PeaceFounder.Server.Controllers.BallotBoxController — Type

mutable struct BallotBoxController
    ledger::BallotBoxLedger
    voters::Set{Pseudonym} # better than members
    collector::Pseudonym
    seed::Union{Digest, Nothing}
    queue::Vector{Vote}
    tree::HistoryTree
    commit::Union{Commit{BallotBoxState}, Nothing}
end

Represents a ballot box for a proposal. Contains proposal, a set of eligiable voters a collector who collects the votes and a seed which is selected at random when the voting starts. queue contains a list of valid votes which yet to be comitted to a ledger. A history tree is built on leafs of ledger's receipts (see a receipt method). A commit contains a collector seal on the current ballotbox state.

Interface: reset_tree!, generator, uuid, ledger, spine, index, seed, leaf, root, receipt, commit, tally, set_seed!, ack_leaf, ack_root, ack_cast, commit_index, commit_state, push!, state, validate, record!, commit!

PeaceFounder.Server.Controllers.PollingStation — Type

struct PollingStation
    halls::Vector{BallotBoxController}
    crypto::CryptoSpec
end

Represents a pooling station which hosts ballotbox ledgers for every proposal collector manages.

Interface: init!, record!, commit!, commit, ack_leaf, ack_root, ack_cast, receipt, spine, ledger, tally, set_seed!

Base.get — Method

get(station::PollingStation, uuid::UUID)::BallotBoxController

Return a ballotbox ledger with a provided UUID. If none is found throws an error.

Base.get — Method

get(station::PollingStation, proposal::Digest)::BallotBoxController

Return a ballotbox which has proposal with provided digest.

Base.getindex — Method

getindex(ledger::BallotBoxController, index::Int)::CastRecord

Return a ledger record at provided index.

Base.length — Method

length(ledger::BallotBoxController)

Return a total length of the ledger including uncommited records in the queue.

Base.push! — Method

push!(ledger::BallotBoxController, record::CastRecord)

Push a record to the ledger bypassing integrity checks. Used when loading the ledger from a trusted source such as local disk or an archive with a signed root cheksum.

HistoryTrees.leaf — Method

leaf(ledger::BallotBoxController, N::Int)::Digest

Return a record digest used to form a history tree.

HistoryTrees.root — Method

root(ledger::BallotBoxController[, N::Int])::Digest

Calculate a root for history tree at given index N. If index is not specified returns the current value.

PeaceFounder.Core.Model.commit — Method

commit(ledger::BallotBoxController)

Return a commit for the ballotbox ledger.

PeaceFounder.Core.Model.commit — Method

commit(station::PollingStation, uuid::UUID)::Commit

Return a ballotbox commit.

PeaceFounder.Core.Model.generator — Method

generator(ledger::BallotBoxController)

Return a relative generator which members use to sign votes anchored by the proposal.

PeaceFounder.Core.Model.index — Method

index(ledger::BallotBoxController)

Return the current index of the ledger. See also length.

PeaceFounder.Core.Model.isbinding — Method

isbinding(vote::Vote, ack::CastAck, hasher)

Check whether acknowledgment is bound to the provided vote.

PeaceFounder.Core.Model.receipt — Method

receipt(ledger::BallotBoxController, index::Int)::CastReceipt

Return a receipt for a ledger element.

PeaceFounder.Core.Model.receipt — Method

receipt(station::PollingStation, uuid::UUID, N::Int)::CastReceipt

Return a receipt for a record with index N at ballotbox with uuid.

PeaceFounder.Core.Model.seed — Method

seed(ledger::BallotBoxController)::Union{Digest, Nothing}

Return a random selected seed used in the voting.

PeaceFounder.Core.Model.tally — Method

tally(station::PollingStation, uuid::UUID)

Compute a tally from ledger records a ballotbox with uuid.

PeaceFounder.Core.Model.uuid — Method

uuid(ledger::BallotBoxController)

Return a UUID of the proposal.

PeaceFounder.Core.Model.voters — Method

voters(ledger::BallotBoxController)

Return a list of member pseudonyms with which members authetificate their votes.

PeaceFounder.Server.Controllers.ack_cast — Method

ack_cast(ledger::BallotBoxController, index::Int)::CastAck

Compute an acknowledgment for record inclusion at index.

PeaceFounder.Server.Controllers.ack_cast — Method

ack_cast(station::PollingStation, uuid::UUID, N::Int)::CastAck

Return inclusion proof with receipt and current tree commit for a leaf at index N and ballotbox with uuid.

PeaceFounder.Server.Controllers.ack_leaf — Method

ack_leaf(ledger::BallotBoxController, index::Int)::AckInclusion

Compute an inclusion proof ::AckInclusion for record element at given index.

PeaceFounder.Server.Controllers.ack_leaf — Method

ack_leaf(station::PollingStation, uuid::UUID, N::Int)::AckInclusion

Return history tree inclusion proof for a tree leaf at index N in ballotbox with uuid.

PeaceFounder.Server.Controllers.ack_root — Method

ack_root(ledger::BallotBoxController, index::Int)::AckConsistency

Compute a history tree consistency proof at index.

PeaceFounder.Server.Controllers.ack_root — Method

ack_root(station::PollingStation, uuid::UUID, N::Int)::AckConsistency

Return history tree consitency proof tree root at index N in ballotbox with uuid.

PeaceFounder.Server.Controllers.commit! — Method

commit!(ledger::BallotBoxController[, timestamp::DataTime], signer::Signer; with_tally=nothing)

Flushes ballotbox ledger's queue and creates a commit for a current ballotbox ledger state with provided timestamp and signer. A keyword argument with_tally has three values true|false to include or exclude a tally from a commited state and nothing which uses a previous commit preference.

PeaceFounder.Server.Controllers.commit! — Method

commit!(station::PollingStation, uuid::UUID, collector::Signer; with_tally = nothing)

Select a ballotbox with provided uuid and commit it's state with collector.

PeaceFounder.Server.Controllers.commit_index — Method

commit_index(ledger::BallotBoxController)::Union{Index, Nothing}

Index at which commit is issued. See also length and index

PeaceFounder.Server.Controllers.commit_state — Method

commit_state(ledger::BallotBoxController)

Return a committed state for a ballotbox ledger.

PeaceFounder.Server.Controllers.init! — Method

init!(station::PollingStation, proposal::Proposal, voters::Set{Pseudonym}[, collector::Pseudonym])

Creates a new ballotbox for given proposal with provided member pseudonyms at a relative generator anchored in the proposal. A collector is optional and provided only when it differs from one specified in the proposal.

PeaceFounder.Server.Controllers.ledger — Method

ledger(ballotbox::BallotBoxController)::Vector{CastRecord}

Return all records from a ballotbox ledger.

PeaceFounder.Server.Controllers.ledger — Method

ledger(station::PollingStation, uuid::UUID)::Vector{CastRecord}

Return a vector of records from a ballotbox with uuid.

PeaceFounder.Server.Controllers.record! — Method

record!(ledger::BallotBoxController, record::CastRecord)

Check the vote in the record for validity and include that in the ledger directly bypassing queue. This method is useful for replaying and debugging ballotbox ledger state changes. See also record!

PeaceFounder.Server.Controllers.record! — Method

record!(ledger::BallotBoxController, vote::Vote)

Check the vote for validity and pushes it to the queue. Returns an index N at which the vote will be recorded in the ledger. See also push!

PeaceFounder.Server.Controllers.record! — Method

record!(station::PollingStation, uuid::UUID, vote::Vote)::Int

Records a vote in a ballotbox with provided proposal UUID. Throws an error if a ballotbox can't be found.

PeaceFounder.Server.Controllers.record! — Method

record!(station::PollingStation, uuid::UUID, vote::Vote)::Int

Record a vote in a ballotbox found by proposal diggest stored in the vote. Throws an error if a ballotbox can't be found.

PeaceFounder.Server.Controllers.reset_tree! — Method

reset_tree!(ledger::BallotBoxController)

Recompute history tree root and cache from the elements in the ledger. This is a useful for loading the ledger all at once.

PeaceFounder.Server.Controllers.set_seed! — Method

set_seed!(ledger::BallotBoxController, seed::Digest)

Set's a seed of the ballotbox.

PeaceFounder.Server.Controllers.set_seed! — Method

set_seed!(station::PollingStation, uuid::UUID, seed::Digest)

Sets a seed for a ballotbox with provided uuid.

PeaceFounder.Server.Controllers.spine — Method

spine(ledger::BallotBoxController)::Vector{Digest}

Return a history tree leaf vector.

PeaceFounder.Server.Controllers.spine — Method

spine(station::PollingStation, uuid::UUID)::Vector{Digest}

Return a leaf vector for a ballotbox with proposal uuid.

PeaceFounder.Server.Controllers.validate — Method

validate(ledger::BallotBoxController, vote::Vote)

Check that vote can be included in the ballotbox. Is well formed, signed by a member pseudonym and cryptographic signature is valid. Raises error if either of checks fail.