PeaceFounder.Model

Crypto primitives

PeaceFounder.Model.canonicalize — Function

ToDo: a well specified encoding is essential here. Binary tree encoding may suffice here. More fancy approach would be to use a DER encoding. Meanwhile JSON shall be used.

PeaceFounder.Model.isbinding — Function

isbinding(x, y, spec::Hash)::Bool
isbinding(x, y, spec) = isbinding(x, y, hasher(spec)::Hash)

Check binding of two objects x and y. Some general examples:

Check that a document is bound to it's signature.

Check that a record is included in the ledger.

Check that a given object is consistent with a ledger.

checks whether request for a new ticketid is properly authetificated.

PeaceFounder.Model.CryptoSpec — Type

struct CryptoSpec
    hasher::Hash
    group::Spec
    generator::Geenrator
end

Specification of cryptographic parameters which are used for public key cryptography, message hashing and authetification codes.

PeaceFounder.Model.verify — Function

verify(message, seal::Seal, [generator::Generator], crypto::CryptoSpec)::Bool
verify(message, pk::Pseudonym, sig::Signature, [generator::Generator], crypto::CryptoSpec)::Bool

Verify the cryptographic signature of the message returning true if valid. An optional generator can be given when signature is issued on a relative generator differing from a base specification crypto.

verify(document[, generator::Generator], crypto::CryptoSpec)::Bool

Verify a cryptographic signature of the document returning true if valid.

verify(braidwork::BraidWork, crypto::CryptoSpec)::Bool

Verify a braider issued cryptographic signature for the braidwork and a zero knowledge proofs. Returns true if both checks succeed.

PeaceFounder.Model.Pseudonym — Type

struct Pseudonym
    pk::Vector{UInt8}
end

A datatype which stores public key in canonical standart octet form.

PeaceFounder.Model.pseudonym — Function

pseudonym(signer::Signer, [generator])::Pseudonym

Return a pseudonym of a signer at a given relative generator. If generator is not passed returns identity pseudonym. (See also id)

pseudonym(seal::Seal)::Pseudonym

Return a pseudonym of a seal. Note that it is not equal to identity when the signature is issued on a relative generator.

pseudonym(vote::Vote)::Pseudonym

Return a pseudonym used to seal the vote.

PeaceFounder.Model.id — Function

id(document)::Pseudonym

Return identity pseudonym of a document issuer.

id(signer)::Pseudonym

Return identity pseudonym of a signer.

PeaceFounder.Model.issuer — Function

issuer(x)

In case an object x is cryptographically signed return an issuer of who have issued the signature. See also id.

PeaceFounder.Model.Generator — Type

struct Generator
    data::Vector{UInt8}
end

Datatype which stores cryptogrpahic group point in standart octet form intended to be used as a base. See also Pseudonym.

PeaceFounder.Model.generator — Function

generator(spec::Spec)::Generator

Return a generator of spec.

generator(crypto::CryptoSpec)::Generator

Return a generator of the specification.

PeaceFounder.Model.Hash — Type

struct Hash
    spec::String
end

A specification for a hasher. See method digest.

PeaceFounder.Model.hasher — Function

hasher(spec)::Hash

Access a hasher function from a given specification.

PeaceFounder.Model.Digest — Type

struct Digest
    data::Vector{UInt8}
end

A message digest obtained applying a hash function on a message or a document. See method digest.

PeaceFounder.Model.digest — Function

digest(bytes::Vector{UInt8}, hasher::Hash)::Digest
digest(x, spec) = digest(canonicalize(x)::Vector{UInt8}, hasher(spec)::Hash)

Compute a hash digest. When input is not in bytes the canonicalize method is applied first.

digest(message::Vector{UInt8}, hasher::Hash)::Digest
digest(document, spec) = digest(canonicalize(message)::Vector{UInt8}, hasher(spec)::Hash)

Return a resulting digest applying hasher on the given message. When message is not octet string a canonicalize method is applied first.

PeaceFounder.Model.Signer — Type

struct Signer
    spec::CryptoSpec
    pbkey::Pseudonym
    key::BigInt
end

A signer type. See a method generate(Signer, spec) for initialization.

Interface: pseudonym, id, sign, seal, approve

PeaceFounder.Model.sign — Function

sign(message::Vector{UInt8}[, generator::Generator], signer::Signer)::Signature

Sign a bytestring message with signer's private key and specification. When generator is provided it is used as a base for the signature.

sign(digest::Digest[, generator::Generator], signer::Signer)::Signature

Sign a digest as an integer with signer's private key and specification. This method avoids running hashing twice when that is done externally. When generator is provided it is used as a base for the signature.

PeaceFounder.Model.generate — Function

generate(Signer, spec::CryptoSpec)::Signer

Generate a unique private key and return a Signer object.

PeaceFounder.Model.Seal — Type

struct Seal
    pbkey::Pseudonym
    sig::Signature
end

A wrapper type for a signature which adds a public key of signature issuer. See seal method.

Interface: pseudonym, verify

PeaceFounder.Model.seal — Function

seal(message::Vector{UInt8}[, generator::Generator], signer::Signer)::Seal

Sign a bytestring message with signer's private key and specification and return a signature as a Seal. When generator is provided it is used as a base for the signature. See also sign.

PeaceFounder.Model.Commit — Type

Commit{T}
    state::T
    seal::Seal
end

Represents a commited ledger state to which issuer can be held accountable for integrity. It is assumed that T implements index and root necessaary to fix a ledger state.

Interface: id, issuer, verify, index, root, state

PeaceFounder.Model.index — Function

index(x)::Int

Return an index of a ledger state.

index(ack::AckInclusion)::Int

Return an index at which the leaf is recorded in the ledger. To obtain the current ledger index use index(commit(ack)).

index(ack::AckConsistency)

Return an index for a root at which the consistency proof is made. To obtain the current ledger index use index(commit(ack)).

HistoryTrees.root — Function

root(x)::Digest

Return a ledger root hash.

root(x::AckConsistency)

Access a root diggest for which the acknowledgment is made.

PeaceFounder.Model.state — Function

state(commit::Commit{T})::T

Return a ledger state. T implements index and root.

PeaceFounder.Model.AckInclusion — Type

struct AckInclusion{T}
    proof::InclusionProof
    commit::Commit{T}
end

Represents an acknowldgment from the issuer that a leaf is permanently included in the ledger. In case the ledger is tampered with this acknowledgement acts as sufficient proof to blame the issuer.

Interface: leaf, id, issuer, commit, index, verify

HistoryTrees.leaf — Function

leaf(ack::AckInclusion)

Access a leaf diggest for which the acknowledgment is made.

PeaceFounder.Model.AckConsistency — Type

struct AckConsistency{T}
    proof::ConsistencyProof
    commit::Commit{T}
end

Represents an ackknowledgment from the issuer that a root is permanetly included in the ledger. This acknowledgemnt assures that ledger up to index(ack) is included in the current ledger which has has index index(commit(ack)). This is useful in a combination with AckInclusion to privatelly update it's validity rather than asking an explicit element. Also ensures that other elements in the ledger are not being tampered with.

Interface: root, id, issuer, commit, index, verify

PeaceFounder.Model.commit — Function

commit(x)

Access a commit of an object x.

PeaceFounder.Model.HMAC — Type

struct HMAC
    key::Vector{UInt8}
    hasher::Hash
end

Represent a hash message authetification code authorizer.

Interface: hasher, digest, key

PeaceFounder.Model.key — Function

key(x)

Access a secret key of an object x.

Recruitment

PeaceFounder.Model.TicketID — Type

struct TicketID
    id::Vector{UInt8}
end

Represents a unique identifier for which a recruit tooken is issued. In case of necessity id can contain a full document, for instance, registration form, proof of identity and etc. In case a privacy is an issue the id can contain a unique identifier which can be matched to an identity in an external database.