PeaceFounder.Model

PeaceFounder.Model.canonicalize — Function

ToDo: a well specified encoding is essential here. Binary tree encoding may suffice here. More fancy approach would be to use a DER encoding. Meanwhile JSON shall be used.

PeaceFounder.Model.isbinding — Function

isbinding(x, y, [hasher|crypto|deme])::Bool

Checks whether two objects x and y legally binding. Some general examples:

Checking that a document is bound to it's signature without verifying it.

Checking that a record is bound to a state for a ledger.

PeaceFounder.Model.CryptoSpec — Type

struct CryptoSpec
    hasher::Hash
    group::Spec
    generator::Geenrator
end

Specification of cryptographic parameters which are used for public key cryptography, message hashing and authetification codes.

PeaceFounder.Model.verify — Function

verify(message, seal::Seal, [generator::Generator], crypto::CryptoSpec)::Bool
verify(message, pk::Pseudonym, sig::Signature, [generator::Generator], crypto::CryptoSpec)::Bool

checks whether signature of on the given message is cryptographically valid.

verify(document, [generator::Generator], crypto::CryptoSpec)::Bool

checks that signature of the document is cryptographically valid.

verify(braidwork::BraidWork, crypto::CryptoSpec)::Bool

verifies the signatures of the braidwork and as well zero knowledge proofs for the braid.

PeaceFounder.Model.Pseudonym — Type

struct Pseudonym
    pk::Vector{UInt8}
end

datatype which stores public key in canonical standart octet form.

PeaceFounder.Model.pseudonym — Function

pseudonym(signer::Signer, [generator])::Pseudonym

returns a pseudonym for a given genrator. If generator is not passed returns identity pseudonym. (See also id)

pseudonym(seal::Seal)::Pseudonym

returns a pseudonym of a seal. Note that it is not equal to identity when signature is being issued on a relative generator.

pseudonym(vote::Vote)::Pseudonym

returns a pseudonym used to seal the vote.

PeaceFounder.Model.id — Function

id(document)::Pseudonym

returns identity pseudonym of document issuer.

id(signer)::Pseudonym

returns identity pseudonym of a signer.

PeaceFounder.Model.Generator — Type

struct Generator
    data::Vector{UInt8}
end

datatype which stores cryptogrpahic group point in standart octet form intended to be used as a base. See also Pseudonym.

PeaceFounder.Model.generator — Function

generator(spec::Spec)::Generator

returns a standart generator of a given specification.

generator(crypto::CryptoSpec)::Generator

returns a relative generator used in the specification.

PeaceFounder.Model.Digest — Type

struct Digest
    data::Vector{UInt8}
end

a message digest obtained applying a hash function on a message or document.

PeaceFounder.Model.digest — Function

digest(message, hasher::Hash|crypto::CryptoSpec)::Digest

returns a resulting digest applying hasher on the given message. If message is not octet string a canonicalize method is applied first.