PeaceFounder.Model

Primitives

PeaceFounder.Model.AckConsistency — Type

struct AckConsistency{T}
    proof::ConsistencyProof
    commit::Commit{T}
end

Represents an ackknowledgment from the issuer that a root is permanetly included in the ledger. This acknowledgemnt assures that ledger up to index(ack) is included in the current ledger which has has index index(commit(ack)). This is useful in a combination with AckInclusion to privatelly update it's validity rather than asking an explicit element. Also ensures that other elements in the ledger are not being tampered with.

Interface: root, id, issuer, commit, index, verify

PeaceFounder.Model.AckInclusion — Type

struct AckInclusion{T}
    proof::InclusionProof
    commit::Commit{T}
end

Represents an acknowldgment from the issuer that a leaf is permanently included in the ledger. In case the ledger is tampered with this acknowledgement acts as sufficient proof to blame the issuer.

Interface: leaf, id, issuer, commit, index, verify

PeaceFounder.Model.Commit — Type

Commit{T}
    state::T
    seal::Seal
end

Represents a commited ledger state to which issuer can be held accountable for integrity. It is assumed that T implements index and root necessaary to fix a ledger state.

Interface: id, issuer, verify, index, root, state

PeaceFounder.Model.CryptoSpec — Type

struct CryptoSpec
    hasher::HashSpec
    group::GroupSpec
    generator::Generator
end

Specification of cryptographic parameters which are used for public key cryptography, message hashing and authetification codes.

PeaceFounder.Model.Digest — Type

struct Digest
    data::Vector{UInt8}
end

A message digest obtained applying a hash function on a message or a document. See method digest.

PeaceFounder.Model.Generator — Type

struct Generator
    data::Vector{UInt8}
end

Datatype which stores cryptogrpahic group point in standart octet form intended to be used as a base. See also Pseudonym.

PeaceFounder.Model.HMAC — Type

struct HMAC
    key::Vector{UInt8}
    hasher::HashSpec
end

Represent a hash message authetification code authorizer.

Interface: hasher, digest, key

PeaceFounder.Model.HashSpec — Type

struct HashSpec
    spec::String
end

A specification for a hasher. See method digest.

PeaceFounder.Model.Pseudonym — Type

struct Pseudonym
    pk::Vector{UInt8}
end

A datatype which stores public key in canonical standart octet form.

PeaceFounder.Model.Seal — Type

struct Seal
    pbkey::Pseudonym
    sig::Signature
end

A wrapper type for a signature which adds a public key of signature issuer. See seal method.

Interface: pseudonym, verify

PeaceFounder.Model.Signer — Type

struct Signer
    spec::CryptoSpec
    pbkey::Pseudonym
    key::BigInt
end

A signer type. See a method generate(Signer, spec) for initialization.

Interface: pseudonym, id, sign, seal, approve

HistoryTrees.leaf — Method

leaf(ack::AckInclusion)

Access a leaf diggest for which the acknowledgment is made.

HistoryTrees.root — Method

root(x::AckConsistency)

Access a root diggest for which the acknowledgment is made.

HistoryTrees.root — Method

root(x)::Digest

Return a ledger root hash.

PeaceFounder.Model.commit — Method

commit(x)

Access a commit of an object x.

PeaceFounder.Model.digest — Method

digest(message::Vector{UInt8}, hasher::HashSpec)::Digest
digest(document, spec) = digest(canonicalize(message)::Vector{UInt8}, hasher(spec)::HashSpec)

Return a resulting digest applying hasher on the given message. When message is not octet string a canonicalize method is applied first.

PeaceFounder.Model.digest — Method

digest(bytes::Vector{UInt8}, hasher::HashSpec)::Digest
digest(x, spec) = digest(canonicalize(x)::Vector{UInt8}, hasher(spec)::HashSpec)

Compute a hash digest. When input is not in bytes the canonicalize method is applied first.

PeaceFounder.Model.generate — Method

generate(Signer, spec::CryptoSpec)::Signer

Generate a unique private key and return a Signer object.

PeaceFounder.Model.generator — Method

generator(crypto::CryptoSpec)::Generator

Return a generator of the specification.

PeaceFounder.Model.generator — Method

generator(spec::Spec)::Generator

Return a generator of spec.

PeaceFounder.Model.hasher — Method

hasher(spec)::HashSpec

Access a hasher function from a given specification.

PeaceFounder.Model.index — Method

index(ack::AckConsistency)

Return an index for a root at which the consistency proof is made. To obtain the current ledger index use index(commit(ack)).

PeaceFounder.Model.index — Method

index(ack::AckInclusion)::Int

Return an index at which the leaf is recorded in the ledger. To obtain the current ledger index use index(commit(ack)).

PeaceFounder.Model.index — Method

index(x)::Int

Return an index of a ledger state.

PeaceFounder.Model.isbinding — Method

isbinding(x, y, spec::HashSpec)::Bool
isbinding(x, y, spec) = isbinding(x, y, hasher(spec)::HashSpec)

Check binding of two objects x and y. Some general examples:

Check that a document is bound to it's signature.

Check that a record is included in the ledger.

Check that a given object is consistent with a ledger.

PeaceFounder.Model.issuer — Method

issuer(x)

In case an object x is cryptographically signed return an issuer of who have issued the signature. See also id.

PeaceFounder.Model.key — Method

key(x)

Access a secret key of an object x.

PeaceFounder.Model.seal — Method

seal(message::Vector{UInt8}[, generator::Generator], signer::Signer)::Seal

Sign a bytestring message with signer's private key and specification and return a signature as a Seal. When generator is provided it is used as a base for the signature. See also sign.

PeaceFounder.Model.sign — Method

sign(digest::Digest[, generator::Generator], signer::Signer)::Signature

Sign a digest as an integer with signer's private key and specification. This method avoids running hashing twice when that is done externally. When generator is provided it is used as a base for the signature.

PeaceFounder.Model.sign — Method

sign(message::Vector{UInt8}[, generator::Generator], signer::Signer)::Signature

Sign a bytestring message with signer's private key and specification. When generator is provided it is used as a base for the signature.

PeaceFounder.Model.state — Method

state(commit::Commit{T})::T

Return a ledger state. T implements index and root.

PeaceFounder.Model.canonicalize — Function

ToDo: a well specified encoding is essential here. Binary tree encoding may suffice here. More fancy approach would be to use a DER encoding. Meanwhile JSON shall be used.

PeaceFounder.Model.pseudonym — Function

pseudonym(signer::Signer, [generator])::Pseudonym

Return a pseudonym of a signer at a given relative generator. If generator is not passed returns identity pseudonym. (See also id)

pseudonym(seal::Seal)::Pseudonym

Return a pseudonym of a seal. Note that it is not equal to identity when the signature is issued on a relative generator.

pseudonym(vote::Vote)::Pseudonym

Return a pseudonym used to seal the vote.

PeaceFounder.Model.id — Function

id(document)::Pseudonym

Return identity pseudonym of a document issuer.

id(signer)::Pseudonym

Return identity pseudonym of a signer.

PeaceFounder.Model.verify — Function

verify(message, seal::Seal, [generator::Generator], crypto::CryptoSpec)::Bool
verify(message, pk::Pseudonym, sig::Signature, [generator::Generator], crypto::CryptoSpec)::Bool

Verify the cryptographic signature of the message returning true if valid. An optional generator can be given when signature is issued on a relative generator differing from a base specification crypto.

verify(document[, generator::Generator], crypto::CryptoSpec)::Bool

Verify a cryptographic signature of the document returning true if valid.

verify(braidwork::BraidReceipt, crypto::CryptoSpec)::Bool

Verify a braider issued cryptographic signature for the braidwork and a zero knowledge proofs. Returns true if both checks succeed.

Admission

PeaceFounder.Model.Registrar — Type

struct Registrar
    metadata::Ref{Vector{UInt8}} 
    tickets::Vector{Ticket}
    signer::Signer
    hmac::HMAC
end

Represents a state for token registrar service. To initialize the service it's necessary to create a signer who can issue a valid admisssion certificates and a secret key with which a recruit client can exchange authorized messages. See also method generate(Registrar, spec).

Metadata is used as means to securelly deliver to the client most recent server specification.

Interface: select, hmac, hasher, key, id, tickets, in, enlist!, admit!, isadmitted, ticket_status

PeaceFounder.Model.Ticket — Type

mutable struct Ticket
    const ticketid::TicketID
    timestamp::DateTime
    attempt::UInt8
    token::Digest
    tokenid::String # 
    admission::Union{Admission, Nothing}
end

Represents a ticket state for ticket with ticketid. timestamp represents time when the ticket have been issued by a registrar client in authorization system of choice, for instance, Registrars.jl; attempt is a counter with which token can be reset which is calculated with token(ticketid, attempt, hmac). Lastly admission contains a certified member pseudonym which was authetificated by the user with token.

PeaceFounder.Model.TicketStatus — Type

struct TicketStatus ticketid::TicketID timestamp::DateTime admission::Union{Nothing, Admission} end

Represents a public state of a ticket. See ticket_status and isadmitted methods.

Base.in — Method

in(ticketid::TicketID, registrar::Registrar)::Bool

Return true if there already is a ticket with ticketid.

PeaceFounder.Model.admit! — Method

admit!(registrar::Registrar, id::Pseudonym, ticketid::TicketID)::Admission

Attempt to admit an identity pseudonym id for ticket ticketid. Authorization is expected to happen at the service layer using provided token in the invite. If a ticket is already registered return admission if it matches the provided id. Otherwise throe an error.

PeaceFounder.Model.enlist! — Method

enlist!(registrar::Registrar, ticketid::TicketID, timestamp::DateTime; route::URI=registrar.route)::Invite

Registers a new ticket with given TicketID and returns an invite. If ticket is already registered the same invite is returned. Throws an error when ticket is already registered.

PeaceFounder.Model.generate — Method

generate(Registrar, spec::CryptoSpec)

Generate a new token registrar with unique signer and athorization key.

PeaceFounder.Model.hmac — Method

hmac(x)::HMAC

Return HMAC authorizer from a given object.

PeaceFounder.Model.isadmitted — Method

isadmitted(ticketid::TicketID, registrar::Registrar)

Check whether a ticket is already admitted. Returns false when either ticket is nonexistent or it's admission is nothing.

PeaceFounder.Model.isadmitted — Method

isadmitted(status::TicketStatus)

Check whether ticket is addmitted.

PeaceFounder.Model.select — Method

select(Admission, ticketid::TicketID, registrar::Registrar)::Union{Admission, Nothing}

Return admission for a ticket with given a given identity pseudonym from registrar. If no ticket with given id is found OR ticket is not yet admitted returns nothing.

PeaceFounder.Model.select — Method

select(Admission, ticketid::TicketID, registrar::Registrar)::Union{Admission, Nothing}

Return admission for a ticket with given ticketid from registrar. If no ticket with given ticketid is found OR ticket is not yet admitted returns nothing.

PeaceFounder.Model.select — Method

select(T, predicate::Function, registrar::Registrar)::Union{T, Nothing}

From a list of all registrar tickets return T <: Union{Ticket, Admission} for which predicate is true. If none succeds returns nothing.

PeaceFounder.Model.set_demehash! — Method

set_demespec!(registrar::Registrar, spec::Union{Digest, DemeSpec})

Replace metadata for a registrar. Note when data is replaced all unfinalized tokens need to be flushed.

PeaceFounder.Model.ticket_status — Method

ticket_status(ticketid::TicketID, registrar::Registrar)::Union{TicketStatus, Nothing}

Return a ticket status for a ticketid. In case ticket is not found return nothing.

PeaceFounder.Model.tickets — Method

tickets(registrar::Registrar)::Vector{TicketID}

Return a list of registered ticket ids.

PeaceFounder.Model.token — Method

token(ticketid::TicketID, hmac::HMAC)

Compute a recruit token for a given ticketid. Calculates it as token=Hash(Hash(0|key)|attempt|ticketid) where attempt is a counter for which token is issued.

Note: the token generation from key is made in order to support it's local computation on a remote server where QR code for registration is shown within organization website.

BraidChain

PeaceFounder.Model.Admission — Type

struct Admission
    ticketid::TicketID
    id::Pseudonym
    timestamp::DateTime
    approval::Union{Seal, Nothing}
end

Represents an admission certificate for a pseudonym id.

Interface: approve, issuer, id, ticket, [isadmitted]

PeaceFounder.Model.BraidChain — Type

struct BraidChain
    members::Set{Pseudonym}
    ledger::Vector{Transaction}
    spec::DemeSpec
    generator::Generator
    tree::HistoryTree
    commit::Union{Commit{ChainState}, Nothing}
end

Represents a braidchain ledger with it's associated state. Can be instantitated with a demespec file using BraidChain(::DemeSpec) method.

Interface: push!, record!, state, length, list, select, roll, constituents, generator, commit, commit_index, ledger, leaf, root, ack_leaf, ack_root, members, commit!

PeaceFounder.Model.ChainState — Type

struct ChainState
    index::Int
    root::Digest
    generator::Generator
    member_count::Int
end

Represents a chain state metadata which is sufficient for integrity checks.

Interface: index, root, generator

PeaceFounder.Model.DemeSpec — Type

struct DemeSpec <: Transaction
    uuid::UUID
    title::String
    crypto::CryptoSpec
    guardian::Pseudonym
    recorder::Pseudonym
    registrar::Pseudonym
    braider::Pseudonym
    proposer::Pseudonym 
    collector::Pseudonym
    timestamp::Union{DateTime, Nothing} = nothing
    signature::Union{Signature, Nothing} = nothing
end

Represents a deme configuration parameters issued by the guardian.

uuid::UUID an unique random generated community identifier;

title::String a community name with which deme is represented;

crypto::CryptoSpec cryptographic parameters for the deme;

guardian::Pseudonym an issuer for this demespec file. Has authorithy to set a roster:

recorder::Pseudonym an authorithy which has rights to add new transactions and is responsable for braidchain's ledger integrity. Issues Commit{ChainState};

registrar::Pseudonym an authorithy which has rights to authorize new admissions to the deme. See Admission and Membership;

braider::Pseudonym an authorithy which can do a legitimate braid jobs for other demes. See BraidReceipt;

proposer::Pseudonym an authorithy which has rights to issue a proposals for the braidchain. See Proposal;

collector::Pseudonym an authorithy which is repsonsable for collecting votes for proposals. This is also recorded in the proposal itself.

timestamp::Union{DateTime, Nothing} time when signature is being issued;

signature::Union{Signature, Nothing} a guardian issued signature.

PeaceFounder.Model.Membership — Type

struct Membership <: Transaction
    admission::Admission
    generator::Generator
    pseudonym::Pseudonym
    approval::Union{Signature, Nothing} 
end

A new member certificate which rolls in (anouances) it's pseudonym at current generator signed with identity pseudonym certified with admission certificate issued by registrar. This two step process is necessary as a checkpoint in situations when braidchain ledger get's locked during a new member resgistration procedure.

PeaceFounder.Model.TicketID — Type

struct TicketID
    id::Vector{UInt8}
end

Represents a unique identifier for which a recruit tooken is issued. In case of necessity id can contain a full document, for instance, registration form, proof of identity and etc. In case a privacy is an issue the id can contain a unique identifier which can be matched to an identity in an external database.

PeaceFounder.Model.Transaction — Type

Transaction

Represents an abstract record type which can be stored in the braidchain ledger.

Base.push! — Method

push!(ledger::BraidChain, t::Transaction)

Add an element to the BraidChain bypassing transaction verification with the chain. This should only be used when the ledger is loaded from a trusted source like a local disk or when final root hash is validated with a trusted source.

HistoryTrees.leaf — Method

leaf(ledger::BraidChain, N::Int)::Digest

Return a ledger's element digest at given index.

HistoryTrees.root — Method

root(ledger::BraidChain[, N::Int])::Digest

Return a ledger root digest. In case when index is not given a current index is used.

PeaceFounder.Model.ack_leaf — Method

ack_leaf(ledger::BraidChain, index::Int)::AckInclusion

Return a proof for record inclusion with respect to a current braidchain ledger history tree root.

PeaceFounder.Model.ack_root — Method

ack_root(ledger::BraidChain, index::Int)

Return a proof for the ledger root at given index with respect to the current braidchain ledger history tree root.

PeaceFounder.Model.approve — Method

approve(x::T, signer::Signer)::T

Cryptographically sign a document x::T and returns a signed document with the same type. To check whether a document is signed see issuer method.

PeaceFounder.Model.approve — Method

approve(member::Membership, signer::Signer)::Membership

Sign a member certificate and return it with approval field filled.

PeaceFounder.Model.commit! — Method

commit!(ledger::BraidChain, signer::Signer)

Commit a current braidchain ledger state with a signer's issued cryptographic signature.

PeaceFounder.Model.commit — Method

commit(ledger::BraidChain)

Return a current commit for a braichain.

PeaceFounder.Model.constituents — Method

constituents(ledger::BraidChain)::Set{Pseudonym}

Return all member identity pseudonyms.

PeaceFounder.Model.generator — Method

generator(ledger[, index])

Return a generator at braidchain ledger row index. If index is omitted return the current state value.

PeaceFounder.Model.generator — Method

generator(ledger::BraidChain)

Return a current relative generator for a braidchain ledger.

PeaceFounder.Model.generator — Method

generator(member::Membership)::Generator

Generator at which member tries to roll in the braidchain.

PeaceFounder.Model.id — Method

id(member::Membership)::Pseudonym

Identity pseudonym for a member.

PeaceFounder.Model.isbinding — Method

isbinding(admission::Admission, spec::DemeSpec)

Check whether issuer of admission is a registrar set in spec.

PeaceFounder.Model.isbinding — Method

isbinding(record::Transaction, ack::AckInclusion{ChainState}, crypto::CryptoSpec)

A generic method checking whether transaction is included in the braidchain.

PeaceFounder.Model.issuer — Method

issuer(member::Membership)::Pseudonym

The identiy of registrar who signed admission.

PeaceFounder.Model.list — Method

list(T, ledger::BraidChain)::Vector{Tuple{Int, T}}

List braidchain elements with a given type together with their index.

PeaceFounder.Model.members — Method

members(ledger::BraidChain[, index::Int])::Set{Pseudonym}

Return a set of member pseudonyms at relative generator at braidchain ledger row index. If index is omitted return a current state value.

PeaceFounder.Model.pseudonym — Method

pseudonym(member::Membership)::Pseudonym

Pseudonym for a member at the generator(member).

PeaceFounder.Model.reset_tree! — Method

reset_tree!(ledger::BraidChain)

Recompute a chain tree hash.

PeaceFounder.Model.roll — Method

roll(ledger::BraidChain)::Vector{Membership}

Return all member certificates from a braidchain ledger.

PeaceFounder.Model.select — Method

select(T, predicate::Function, ledger::BraidChain)::Union{T, Nothing}

Return a first element from a ledger with a type T which satisfies a predicate.

PeaceFounder.Model.state — Method

state(ledger::BraidChain)

Return a current braidchain ledger state metadata.

PeaceFounder.Model.ticket — Method

ticket(x::Admission)

Return a TicketID which is admitted.

PeaceFounder.Model.ticket — Method

ticket(member::Membership)

Ticket for a member admission certificate.

PeaceFounder.Model.BraidReceipt — Type

struct BraidReceipt <: Transaction
    braid::Simulator
    consumer::DemeSpec
    producer::DemeSpec
    approval::Union{Seal, Nothing}
end

Represents a braider's computation which is supported with zero knowledge proof of shuffle and decryption assuring it's corectness stored in a braid field; consumer denotes a deme for which the braid is intended and producer denotes a deme where the braid is made. To assert latter the the braider signs the braidwork and stores that in the aproval field. See a braid method.

Interface: approve, verify, input_generator, input_members, output_generator, output_members

PeaceFounder.Model.approve — Method

approve(braid::BraidReceipt, braider::Signer)

Sign a braidwork with a braider. Throws an error if braider is not in the producer demespec.

PeaceFounder.Model.braid — Method

braid(generator::Generator, members::Union{Vector{Pseudonym}, Set{Pseudonym}}, consumer::DemeSpec, producer::DemeSpec; verifier = (g) -> ProtocolSpec(; g))

Selects a private exponent x at random and computes a new generator $g' = g^x$ and $member_i'=member_i^x$ returns the latter in a sorted order and provides a zero knowledge proof that all operations have been performed honestly. In partucular, not including/droping new member pseudonyms in the process. consumer attributes are necessary to interepret generator and pseudonym group elements with which the computation is performed.

By default a Verificatum compatable verifier is used for performing reencryption proof of shuffle

A verifier can be configured with a keyword argument. By default a Verificatum compatable verifier for a proof of shuffle is used.

PeaceFounder.Model.input_generator — Method

input_generator(braid::BraidReceipt)

Return input generator of the braid.

PeaceFounder.Model.input_members — Method

input_members(braid::BraidReceipt)

Return input member pseudonyms of the braid at provided input generator. See input_generator

PeaceFounder.Model.output_generator — Method

output_generator(braid::BraidReceipt)

Return output genertor of the braid.

PeaceFounder.Model.output_members — Method

output_members(braid::BraidReceipt)

Return output member pseudonyms of the braid at a resulting output generator. See output_generator

PeaceFounder.Model.verify — Method

verify(braid::BraidReceipt, crypto::CryptoSpec)

Verifies a braid approval and then it's zero knowledge proofs. A crypto argument is provided to avoid downgrading attacks.

Ballot Box

PeaceFounder.Model.Ballot — Type

struct Ballot
    options::Vector{String}
end

Represents a simple ballot form for multiple choice question.

PeaceFounder.Model.BallotBox — Type

mutable struct BallotBox
    proposal::Proposal
    voters::Set{Pseudonym} # better than members
    collector::Pseudonym
    seed::Union{Digest, Nothing}
    crypto::CryptoSpec # on the other hand the inclusion of a vote should be binding enough as it includes proposal hash.
    queue::Vector{Vote}
    ledger::Vector{CastRecord}
    tree::HistoryTree
    commit::Union{Commit{BallotBoxState}, Nothing}
end

Represents a ballot box for a proposal. Contains proposal, a set of eligiable voters a collector who collects the votes and a seed which is selected at random when the voting starts. queue contains a list of valid votes which yet to be comitted to a ledger. A history tree is built on leafs of ledger's receipts (see a receipt method). A commit contains a collector seal on the current ballotbox state.

Interface: reset_tree!, generator, uuid, members, ledger, spine, index, seed, leaf, root, record, receipt, commit, tally, set_seed!, ack_leaf, ack_root, ack_cast, commit_index, commit_state, push!, state, validate, record!, commit!

PeaceFounder.Model.BallotBoxState — Type

struct BallotBoxState
    proposal::Digest
    seed::Digest
    index::Int
    root::Digest
    tally::Union{Nothing, Tally} 
    view::Union{Nothing, BitVector} # 
end

Represents a public ballot box state. Contains an immutable proposal and seed digest; a current ledger index, history tree root. When ellections end a tally is included in the state and a view is added listing all counted votes. Note that the view attribute is important for a client to know whether it's key have leaked and somone lese havbe superseeded it's vote by revoting.

PeaceFounder.Model.CastAck — Type

struct CastAck
    receipt::CastReceipt
    ack::AckInclusion{BallotBoxState}
end

Represents a reply to a voter when a vote have been included in the ballotbox ledger. Contains a receipt and inlcusion acknowledgment. In future would also include a blind signature in the reply for a proof of participation. To receive this reply (with a blind signature in the future) a voter needs to send a vote which is concealed during ellections and thus tagging votes with blind signatures from reply to monitor revoting would be as hard as monitoring the submitted votes.

PeaceFounder.Model.CastReceipt — Type

struct CastReceipt
    vote::Digest
    timestamp::DateTime
end

Represents a ballotbox ledger receipt which is hashed for a history tree. It's sole purpose is to assure voters that their vote is included in the ledger while also adding additional metadata as timestamp and. In contrast to CastRecord it does not reveal how voter have voted thus can be published during ellections without violating fairness property. For some situations it may be useful to extend the time until the votes are published as that can disincentivice coercers and bribers as they would not know whether their coerced vote have been superseeded in revoting. See receipt method for it's construction from a CastRecord.

Note that a blind signature could be commited as H(signature|H(vote)) to avoid tagging the use of pseudonym during ellections while collector could issue only a one blind signature for a voter. Note that members whoose private key could have been stolen could not obtain a valid signature for participation and that could be a good thing!

PeaceFounder.Model.CastRecord — Type

struct CastRecord
    vote::Vote
    timestamp::DateTime
end

Represents a ballotbox ledger record. Adds a timestamp when the vote have been recorded. In future, the record will also contain a blind signature with which members could prove to everyone that they had cast their vote without revealing the link to the vote.

PeaceFounder.Model.PollingStation — Type

struct PollingStation
    halls::Vector{BallotBox}
    crypto::CryptoSpec
end

Represents a pooling station which hosts ballotbox ledgers for every proposal collector manages.

Interface: add!, ballotbox, record!, commit!, commit, ack_leaf, ack_root, ack_cast, record, receipt, spine, ledger, tally, set_seed!

PeaceFounder.Model.Proposal — Type

struct Proposal <: Transaction 
    uuid::UUID
    summary::String
    description::String
    ballot::Ballot
    open::DateTime
    closed::DateTime
    collector::Union{Pseudonym, Nothing} # 
    anchor::Union{ChainState, Nothing}
    approval::Union{Seal, Nothing} 
end

Represents a proposal for a ballot specyfing voting window, unique identifier, summary, description. Set's the collector identity which collects votes and issues vote inclusion receipts and is responsable for maintaining the ledger's integrity. The proposal also includes an anchor which sets a relative generator with which members vote anonymously. To be considered valid is signed by proposer authorizing vote to take place.

PeaceFounder.Model.Selection — Type

struct Selection
    option::Int
end

Represents voter's selection for a Ballot form.

PeaceFounder.Model.Tally — Type

struct Tally
    data::Vector{Int}
end

Represent a tally for Ballot form obtained aftert counting multiple voter's Selection forms.

PeaceFounder.Model.Vote — Type

struct Vote
    proposal::Digest
    seed::Digest
    selection::Selection
    seq::Int
    approval::Union{Seal, Nothing} 
end

Represents a vote for a proposal issued by a member. The proposal is stored as hash digest ensuring that member have voted on an untampered proposal. seed contains a randon string issued by collector at the moment when a vote starts to eliminate early voting / shortening a time at which coercers could act uppon. selection contians voter's preference;

seq is a serquence number counting a number of votes at which vote have been approved for a given proposal. It starts at 1 and is increased by one for every single signature made on the proposal. This is an important measure which allows to detect possible leakage of a member's private key. Also provides means for revoting ensuring that latest vote get's counted.

The vote is considered valid when it is sealed by a member's private key at a relative generator stored in the proposal.

Base.length — Method

length(ledger::BallotBox)

Return a total length of the ledger including uncommited records in the queue.

Base.push! — Method

push!(ledger::BallotBox, record::CastRecord)

Push a record to the ledger bypassing integrity checks. Used when loading the ledger from a trusted source such as local disk or an archive with a signed root cheksum.

HistoryTrees.leaf — Method

leaf(ledger::BallotBox, N::Int)::Digest

Return a record digest used to form a history tree.

HistoryTrees.root — Method

root(ledger::BallotBox[, N::Int])::Digest

Calculate a root for history tree at given index N. If index is not specified returns the current value.

HistoryTrees.root — Method

root(state::BallotBoxState)

Return a history tree root for a current ballotbox ledger state.

PeaceFounder.Model.ack_cast — Method

ack_cast(ledger::BallotBox, index::Int)::CastAck

Compute an acknowledgment for record inclusion at index.

PeaceFounder.Model.ack_cast — Method

ack_cast(station::PollingStation, uuid::UUID, N::Int)::CastAck

Return inclusion proof with receipt and current tree commit for a leaf at index N and ballotbox with uuid.

PeaceFounder.Model.ack_leaf — Method

ack_leaf(ledger::BallotBox, index::Int)::AckInclusion

Compute an inclusion proof ::AckInclusion for record element at given index.

PeaceFounder.Model.ack_leaf — Method

ack_leaf(station::PollingStation, uuid::UUID, N::Int)::AckInclusion

Return history tree inclusion proof for a tree leaf at index N in ballotbox with uuid.

PeaceFounder.Model.ack_root — Method

ack_root(ledger::BallotBox, index::Int)::AckConsistency

Compute a history tree consistency proof at index.

PeaceFounder.Model.ack_root — Method

ack_root(station::PollingStation, uuid::UUID, N::Int)::AckConsistency

Return history tree consitency proof tree root at index N in ballotbox with uuid.

PeaceFounder.Model.add! — Method

add!(station::PollingStation, proposal::Proposal, voters::Set{Pseudonym}[, collector::Pseudonym])

Creates a new ballotbox for given proposal with provided member pseudonyms at a relative generator anchored in the proposal. A collector is optional and provided only when it differs from one specified in the proposal.

PeaceFounder.Model.ballotbox — Method

ballotbox(station::PollingStation, uuid::UUID)::BallotBox

Return a ballotbox ledger with a provided UUID. If none is found throws an error.

PeaceFounder.Model.ballotbox — Method

ballotbox(station::PollingStation, proposal::Digest)::BallotBox

Return a ballotbox which has proposal with provided digest.

PeaceFounder.Model.commit! — Method

commit!(ledger::BallotBox[, timestamp::DataTime], signer::Signer; with_tally=nothing)

Flushes ballotbox ledger's queue and creates a commit for a current ballotbox ledger state with provided timestamp and signer. A keyword argument with_tally has three values true|false to include or exclude a tally from a commited state and nothing which uses a previous commit preference.

PeaceFounder.Model.commit! — Method

commit!(station::PollingStation, uuid::UUID, collector::Signer; with_tally = nothing)

Select a ballotbox with provided uuid and commit it's state with collector.

PeaceFounder.Model.commit — Method

commit(ledger::BallotBox)

Return a commit for the ballotbox ledger.

PeaceFounder.Model.commit — Method

commit(ack::CastAck)

Return a commit from a CastAck.

PeaceFounder.Model.commit — Method

commit(station::PollingStation, uuid::UUID)::Commit

Return a ballotbox commit.

PeaceFounder.Model.commit_index — Method

commit_index(ledger::BallotBox)::Union{Index, Nothing}

Index at which commit is issued. See also length and index

PeaceFounder.Model.commit_state — Method

commit_state(ledger::BallotBox)

Return a committed state for a ballotbox ledger.

PeaceFounder.Model.generator — Method

generator(ledger::BallotBox)

Return a relative generator which members use to sign votes anchored by the proposal.

PeaceFounder.Model.generator — Method

generator(proposal::Proposal)

A relative generator for at which memebers sign votes for this proposal.

PeaceFounder.Model.index — Method

index(state::BallotBoxState)

Return an index for a current ballotbox ledger state.

PeaceFounder.Model.index — Method

index(ledger::BallotBox)

Return the current index of the ledger. See also length.

PeaceFounder.Model.isbinding — Method

isbinding(chain::BraidChain, state::ChainState)::Bool

Check that chain state is consistent with braidchain ledger.

PeaceFounder.Model.isbinding — Method

isbinding(ack::CastAck, proposal::Proposal, hasher::HashSpec)::Bool

Check that acknowledgment is legitimate meaning that it is issued by a collector listed in the proposal.

PeaceFounder.Model.isbinding — Method

isbinding(receipt::CastReceipt, ack::AckInclusion, hasher::HashSpec)::Bool

Check that cast receipt is binding to received inclusion acknowledgment.

PeaceFounder.Model.isbinding — Method

isbinding(receipt::CastReceipt, vote::Vote, hasher::HashSpec)::Bool

Check that the receipt is bidning to a vote.

PeaceFounder.Model.isbinding — Method

isbinding(vote::Vote, ack::CastAck, hasher)

Check whether acknowledgment is bound to the provided vote.

PeaceFounder.Model.isbinding — Method

isbinding(vote::Vote, proposal::Proposal, crypto::CryptoSpec)

Check that the vote is bound to a proposal..

PeaceFounder.Model.isconsistent — Method

isconsistent(selection::Selection, ballot::Ballot)

Verifies that voter's selection is consistent with ballot form. For instance, whether selection is withing the range of ballot options.

PeaceFounder.Model.issuer — Method

issuer(proposal::Proposal)

Issuer of approval for the proposal.

PeaceFounder.Model.ledger — Method

ledger(ballotbox::BallotBox)::Vector{CastRecord}

Return all records from a ballotbox ledger.

PeaceFounder.Model.ledger — Method

ledger(station::PollingStation, uuid::UUID)::Vector{CastRecord}

Return a vector of records from a ballotbox with uuid.

PeaceFounder.Model.pseudonym — Method

pseudonym(vote::Vote)::Union{Pseudonym, Nothing}

Return a pseudonym with which vote is sealed.

PeaceFounder.Model.receipt — Method

receipt(ledger::BallotBox, index::Int)::CastReceipt

Return a receipt for a ledger element.

PeaceFounder.Model.receipt — Method

receipt(record::CastRecord, hasher::HashSpec)::CastReceipt

Construct a CastReceipt from a CastRecord with a provided hasher function.

PeaceFounder.Model.receipt — Method

receipt(station::PollingStation, uuid::UUID, N::Int)::CastReceipt

Return a receipt for a record with index N at ballotbox with uuid.

PeaceFounder.Model.record! — Method

record!(ledger::BallotBox, record::CastRecord)

Check the vote in the record for validity and include that in the ledger directly bypassing queue. This method is useful for replaying and debugging ballotbox ledger state changes. See also record!

PeaceFounder.Model.record! — Method

record!(ledger::BallotBox, vote::Vote)

Check the vote for validity and pushes it to the queue. Returns an index N at which the vote will be recorded in the ledger. See also push!

PeaceFounder.Model.record! — Method

record!(station::PollingStation, uuid::UUID, vote::Vote)::Int

Records a vote in a ballotbox with provided proposal UUID. Throws an error if a ballotbox can't be found.

PeaceFounder.Model.record! — Method

record!(station::PollingStation, uuid::UUID, vote::Vote)::Int

Record a vote in a ballotbox found by proposal diggest stored in the vote. Throws an error if a ballotbox can't be found.

PeaceFounder.Model.record — Method

record(ledger::BallotBox, index::Int)::CastRecord

Return a ledger record at provided index.

PeaceFounder.Model.record — Method

record(station::PollingStation, uuid::UUID, N::Int)::CastRecord

Return a record with an index N at ballotbox with uuid.

PeaceFounder.Model.reset_tree! — Method

reset_tree!(ledger::BallotBox)

Recompute history tree root and cache from the elements in the ledger. This is a useful for loading the ledger all at once.

PeaceFounder.Model.seed — Method

seed(ledger::BallotBox)::Union{Digest, Nothing}

Return a random selected seed used in the voting.

PeaceFounder.Model.set_seed! — Method

set_seed!(ledger::BallotBox, seed::Digest)

Set's a seed of the ballotbox.

PeaceFounder.Model.set_seed! — Method

set_seed!(station::PollingStation, uuid::UUID, seed::Digest)

Sets a seed for a ballotbox with provided uuid.

PeaceFounder.Model.spine — Method

spine(ledger::BallotBox)::Vector{Digest}

Return a history tree leaf vector.

PeaceFounder.Model.spine — Method

spine(station::PollingStation, uuid::UUID)::Vector{Digest}

Return a leaf vector for a ballotbox with proposal uuid.

PeaceFounder.Model.state — Method

state(ledger::BallotBox; with_tally::Union{Nothing, Bool} = nothing)::BallotBoxState

Return a state metadata for ballotbox ledger.

PeaceFounder.Model.state — Method

state(proposal::Proposal)::ChainState

A braidchain ledger state which is used to anchor a relative generator for members.

PeaceFounder.Model.tally — Method

tally(ballot::Ballot, ballots::AbstractVector{Selection})::Tally

Count ballots, check that they are filled consistently and return a final tally.

PeaceFounder.Model.tally — Method

tally(ledger::BallotBox)

Compute a tally for a ballotbox ledger.

PeaceFounder.Model.tally — Method

tally(station::PollingStation, uuid::UUID)

Compute a tally from ledger records a ballotbox with uuid.

PeaceFounder.Model.uuid — Method

uuid(ledger::BallotBox)

Return a UUID of the proposal.

PeaceFounder.Model.uuid — Method

uuid(proposal::Proposal)::UUID

UUID for a proposal. Issued by proposer and it's purpose is to croslink to an external system durring the proposal dreafting stage.

PeaceFounder.Model.validate — Method

validate(ledger::BallotBox, vote::Vote)

Check that vote can be included in the ballotbox. Is well formed, signed by a member pseudonym and cryptographic signature is valid. Raises error if either of checks fail.

PeaceFounder.Model.verify — Method

verify(ack::CastAck, crypto::CryptoSpec)::Bool

Verify the cast acknowledgment cryptographic signature.

PeaceFounder.Model.vote — Method

vote(proposal::Proposal, seed::Digest, selection::Selection, member::Signer; seq = 1)

Issue a vote on a proposal and provided collector seed for a member's selection.

PeaceFounder.Model.voters — Method

members(ledger::BallotBox)

Return a list of member pseudonyms with which members authetificate their votes.