PeaceFounder.Model

Crypto primitives

PeaceFounder.Model.canonicalize — Function

ToDo: a well specified encoding is essential here. Binary tree encoding may suffice here. More fancy approach would be to use a DER encoding. Meanwhile JSON shall be used.

PeaceFounder.Model.isbinding — Function

isbinding(x, y, spec::Hash)::Bool
isbinding(x, y, spec) = isbinding(x, y, hasher(spec)::Hash)

Check binding of two objects x and y. Some general examples:

Check that a document is bound to it's signature.

Check that a record is included in the ledger.

Check that a given object is consistent with a ledger.

isbinding(ticketid::TicketID, timestamp::DateTime, auth_code::Digest, hmac::HMAC)

Check whether request for a new ticketid is authorized by recruit client. Note that freshness of the request is not considered within this method.

isbinding(metadata::Vector{UInt8}, ticketid::TicketID, salt::Vector{UInt8}, auth_code::Digest, hmac::HMAC)

Check whether a reply for a new ticketid is authorized by the recruit server.

isbinding(id::Pseudonym, auth_code::Digest, token::Digest, hasher::Hash)

Check whether a request for a new identity pseudonym admission is authetificated. The token is delivered to a user by recruit client.

isbinding(record::Transaction, ack::AckInclusion{ChainState}, crypto::CryptoSpec)

A generic method checking whether transaction is included in the braidchain.

isbinding(admission::Admission, spec::DemeSpec)

Check whether issuer of admission is a recruiter set in spec.

PeaceFounder.Model.CryptoSpec — Type

struct CryptoSpec
    hasher::Hash
    group::Spec
    generator::Geenrator
end

Specification of cryptographic parameters which are used for public key cryptography, message hashing and authetification codes.

PeaceFounder.Model.verify — Function

verify(message, seal::Seal, [generator::Generator], crypto::CryptoSpec)::Bool
verify(message, pk::Pseudonym, sig::Signature, [generator::Generator], crypto::CryptoSpec)::Bool

Verify the cryptographic signature of the message returning true if valid. An optional generator can be given when signature is issued on a relative generator differing from a base specification crypto.

verify(document[, generator::Generator], crypto::CryptoSpec)::Bool

Verify a cryptographic signature of the document returning true if valid.

verify(braidwork::BraidWork, crypto::CryptoSpec)::Bool

Verify a braider issued cryptographic signature for the braidwork and a zero knowledge proofs. Returns true if both checks succeed.

PeaceFounder.Model.Pseudonym — Type

struct Pseudonym
    pk::Vector{UInt8}
end

A datatype which stores public key in canonical standart octet form.

PeaceFounder.Model.pseudonym — Function

pseudonym(signer::Signer, [generator])::Pseudonym

Return a pseudonym of a signer at a given relative generator. If generator is not passed returns identity pseudonym. (See also id)

pseudonym(seal::Seal)::Pseudonym

Return a pseudonym of a seal. Note that it is not equal to identity when the signature is issued on a relative generator.

pseudonym(vote::Vote)::Pseudonym

Return a pseudonym used to seal the vote.

PeaceFounder.Model.id — Function

id(document)::Pseudonym

Return identity pseudonym of a document issuer.

id(signer)::Pseudonym

Return identity pseudonym of a signer.

PeaceFounder.Model.issuer — Function

issuer(x)

In case an object x is cryptographically signed return an issuer of who have issued the signature. See also id.

PeaceFounder.Model.Generator — Type

struct Generator
    data::Vector{UInt8}
end

Datatype which stores cryptogrpahic group point in standart octet form intended to be used as a base. See also Pseudonym.

PeaceFounder.Model.generator — Function

generator(spec::Spec)::Generator

Return a generator of spec.

generator(crypto::CryptoSpec)::Generator

Return a generator of the specification.

PeaceFounder.Model.Hash — Type

struct Hash
    spec::String
end

A specification for a hasher. See method digest.

PeaceFounder.Model.hasher — Function

hasher(spec)::Hash

Access a hasher function from a given specification.

PeaceFounder.Model.Digest — Type

struct Digest
    data::Vector{UInt8}
end

A message digest obtained applying a hash function on a message or a document. See method digest.

PeaceFounder.Model.digest — Function

digest(bytes::Vector{UInt8}, hasher::Hash)::Digest
digest(x, spec) = digest(canonicalize(x)::Vector{UInt8}, hasher(spec)::Hash)

Compute a hash digest. When input is not in bytes the canonicalize method is applied first.

digest(message::Vector{UInt8}, hasher::Hash)::Digest
digest(document, spec) = digest(canonicalize(message)::Vector{UInt8}, hasher(spec)::Hash)

Return a resulting digest applying hasher on the given message. When message is not octet string a canonicalize method is applied first.

PeaceFounder.Model.Signer — Type

struct Signer
    spec::CryptoSpec
    pbkey::Pseudonym
    key::BigInt
end

A signer type. See a method generate(Signer, spec) for initialization.

Interface: pseudonym, id, sign, seal, approve

PeaceFounder.Model.sign — Function

sign(message::Vector{UInt8}[, generator::Generator], signer::Signer)::Signature

Sign a bytestring message with signer's private key and specification. When generator is provided it is used as a base for the signature.

sign(digest::Digest[, generator::Generator], signer::Signer)::Signature

Sign a digest as an integer with signer's private key and specification. This method avoids running hashing twice when that is done externally. When generator is provided it is used as a base for the signature.

PeaceFounder.Model.generate — Function

generate(Signer, spec::CryptoSpec)::Signer

Generate a unique private key and return a Signer object.

generate(TokenRecruiter, spec::CryptoSpec)

Generate a new token recruiter with unique signer and athorization key.

PeaceFounder.Model.Seal — Type

struct Seal
    pbkey::Pseudonym
    sig::Signature
end

A wrapper type for a signature which adds a public key of signature issuer. See seal method.

Interface: pseudonym, verify

PeaceFounder.Model.seal — Function

seal(message::Vector{UInt8}[, generator::Generator], signer::Signer)::Seal

Sign a bytestring message with signer's private key and specification and return a signature as a Seal. When generator is provided it is used as a base for the signature. See also sign.

PeaceFounder.Model.approve — Function

approve(x::T, signer::Signer)::T

Cryptographically sign a document x::T and returns a signed document with the same type. To check whether a document is signed see issuer method.

PeaceFounder.Model.Commit — Type

Commit{T}
    state::T
    seal::Seal
end

Represents a commited ledger state to which issuer can be held accountable for integrity. It is assumed that T implements index and root necessaary to fix a ledger state.

Interface: id, issuer, verify, index, root, state

PeaceFounder.Model.index — Function

index(x)::Int

Return an index of a ledger state.

index(ack::AckInclusion)::Int

Return an index at which the leaf is recorded in the ledger. To obtain the current ledger index use index(commit(ack)).

index(ack::AckConsistency)

Return an index for a root at which the consistency proof is made. To obtain the current ledger index use index(commit(ack)).

HistoryTrees.root — Function

root(x)::Digest

Return a ledger root hash.

root(x::AckConsistency)

Access a root diggest for which the acknowledgment is made.

PeaceFounder.Model.state — Function

state(commit::Commit{T})::T

Return a ledger state. T implements index and root.

PeaceFounder.Model.AckInclusion — Type

struct AckInclusion{T}
    proof::InclusionProof
    commit::Commit{T}
end

Represents an acknowldgment from the issuer that a leaf is permanently included in the ledger. In case the ledger is tampered with this acknowledgement acts as sufficient proof to blame the issuer.

Interface: leaf, id, issuer, commit, index, verify

HistoryTrees.leaf — Function

leaf(ack::AckInclusion)

Access a leaf diggest for which the acknowledgment is made.

PeaceFounder.Model.AckConsistency — Type

struct AckConsistency{T}
    proof::ConsistencyProof
    commit::Commit{T}
end

Represents an ackknowledgment from the issuer that a root is permanetly included in the ledger. This acknowledgemnt assures that ledger up to index(ack) is included in the current ledger which has has index index(commit(ack)). This is useful in a combination with AckInclusion to privatelly update it's validity rather than asking an explicit element. Also ensures that other elements in the ledger are not being tampered with.

Interface: root, id, issuer, commit, index, verify

PeaceFounder.Model.commit — Function

commit(x)

Access a commit of an object x.

PeaceFounder.Model.HMAC — Type

struct HMAC
    key::Vector{UInt8}
    hasher::Hash
end

Represent a hash message authetification code authorizer.

Interface: hasher, digest, key

PeaceFounder.Model.key — Function

key(x)

Access a secret key of an object x.

PeaceFounder.Model.hmac

Recruitment

PeaceFounder.Model.TicketID — Type

struct TicketID
    id::Vector{UInt8}
end

Represents a unique identifier for which a recruit tooken is issued. In case of necessity id can contain a full document, for instance, registration form, proof of identity and etc. In case a privacy is an issue the id can contain a unique identifier which can be matched to an identity in an external database.

PeaceFounder.Model.Admission — Type

struct Admission
    ticketid::TicketID
    id::Pseudonym
    timestamp::DateTime
    approval::Union{Seal, Nothing}
end

Represents an admission certificate for a pseudonym id.

Interface: approve, issuer, id, ticket, isadmitted

PeaceFounder.Model.ticket — Function

ticket(x::Admission)

Return a TicketID which is admitted.

PeaceFounder.Model.Ticket — Type

mutable struct Ticket
    const ticketid::TicketID
    timestamp::DateTime
    salt::Vector{UInt8}
    auth_code::Digest
    token::Digest
    admission::Union{Admission, Nothing}
end

Represents a ticket state for ticket with ticketid. timestamp represents time when the ticket have been issued by a recruiter client in authorization system of choice, for instance, Recruiters.jl; salt contains a random bytestring generated by the server from which the reccruiter client can derive a token as token(ticketid, salt, hmac). auth_code is a server generated authetification code for a given salt and current metadata. Lastly admission contains a certified member pseudonym which was authetificated by the user with token.

PeaceFounder.Model.TokenRecruiter — Type

struct TokenRecruiter
    metadata::Ref{Vector{UInt8}} 
    tickets::Vector{Ticket}
    signer::Signer
    hmac::HMAC
end

Represents a state for token recruiter service. To initialize the service it's necessary to create a signer who can issue a valid admisssion certificates and a secret key with which a recruit client can exchange authorized messages. See also method generate(TokenRecruiter, spec).

Metadata is used as means to securelly deliver to the client most recent server specification.

Interface: select, hmac, hasher, key, id, tickets, in, set_metadata!, enlist!, admit!, isadmitted, ticket_status

Base.in — Function

in(ticketid::TicketID, recruiter::TokenRecruiter)::Bool

Return true if there already is a ticket with ticketid.

PeaceFounder.Model.set_metadata! — Function

set_metadata!(recruiter::TokenRecruiter, metadata::Vector{UInt8})

Replace metadata for a recruiter. Note when data is replaced all unfinalized tokens need to be flushed.

PeaceFounder.Model.select — Function

select(T, predicate::Function, recruiter::TokenRecruiter)::Union{T, Nothing}

From a list of all recruiter tickets return T <: Union{Ticket, Admission} for which predicate is true. If none succeds returns nothing.

select(Admission, ticketid::TicketID, recruiter::TokenRecruiter)::Union{Admission, Nothing}

Return admission for a ticket with given ticketid from recruiter. If no ticket with given ticketid is found OR ticket is not yet admitted returns nothing.

select(Admission, ticketid::TicketID, recruiter::TokenRecruiter)::Union{Admission, Nothing}

Return admission for a ticket with given a given identity pseudonym from recruiter. If no ticket with given id is found OR ticket is not yet admitted returns nothing.

PeaceFounder.Model.tickets — Function

tickets(recruiter::TokenRecruiter)::Vector{TicketID}

Return a list of registered ticket ids.

PeaceFounder.Model.auth — Function

auth(ticketid::TicketID, time::DateTime, hmac::HMAC)::Digest

Compute a hash authorization code for ticketid at given time. The message needs to reach token recruiter within a 60 second window before it is obseleted.

auth(metadata::Vector{UInt8}, ticketid::TicketID, salt::Vector{UInt8}, hmac::HMAC)::Digest

Compute a hash authorization code for a metadata, salt from which a token is derived. Metadata can be configured to contain most recent server specification DemeSpec; ticketid takes session identifier role.

auth(id::Pseudonym, hmac::HMAC)::Digest
auth(id::Pseudonym, token::Digest, hasher::Hash) = auth(id, HMAC(bytes(token), hasher))

Compute a hash authorization code for identity pseudonym using token as a key.

PeaceFounder.Model.token — Function

token(ticketid::TicketID, salt::Vector{UInt8}, hmac::HMAC)

Compute a recruit token for a given ticket with provided salt.

PeaceFounder.Model.enlist! — Function

enlist!(recruiter::TokenRecruiter, ticketid::TicketID, timestamp::DateTime, ticket_auth_code::Digest)::Tuple{Vector{UInt8}, Vector{UInt8}, Digest}

Attempt to enlist a ticket with given ticketid authetificated by a recruit client at timestmap. This function checks the age of the request which need to be less than 60 seconds to be considered. Then the hash authorization code is checked after which a triplet of metadata, salt and reply_auth_code is returned. If either of theses checks fail an error is raised and needs to be dealt by the user. See a token method on how the token is derived.

PeaceFounder.Model.admit! — Function

admit!(recruiter::TokenRecruiter, id::Pseudonym, ticketid::TicketID, auth_code::Digest)::Admission

Attempt to admit an identity pseudonym id for ticket ticketid with provided authorization code. This function retrieves a ticket with given ticketid and uses it's recorded token to check whether the request is binding. In the case of success an admnission certificate is formed with provided indenty pseudonym id and is signed by the recruter's private key. Otherwise when either of checks fail an error is raised. In the case ticket is already addmitted, returns previously stored admission.

PeaceFounder.Model.isadmitted — Function

isadmitted(ticketid::TicketID, recruiter::TokenRecruiter)

Check whether a ticket is already admitted. Returns false when either ticket is nonexistent or it's admission is nothing.

isadmitted(status::TicketStatus)

Check whether ticket is addmitted.

PeaceFounder.Model.TicketStatus — Type

struct TicketStatus ticketid::TicketID timestamp::DateTime admission::Union{Nothing, Admission} end

Represents a public state of a ticket. See ticket_status and isadmitted methods.

PeaceFounder.Model.ticket_status — Function

ticket_status(ticketid::TicketID, recruiter::TokenRecruiter)::Union{TicketStatus, Nothing}

Return a ticket status for a ticketid. In case ticket is not found return nothing.

PeaceFounder.Model.hmac — Function

hmac(x)::HMAC

Return HMAC authorizer from a given object.

BraidChain

PeaceFounder.Model.Transaction — Type

Represents an abstract record type which can be stored in the braidchain ledger.

PeaceFounder.Model.DemeSpec — Type

struct DemeSpec <: Transaction
    uuid::UUID
    title::String
    crypto::CryptoSpec
    guardian::Pseudonym
    recorder::Pseudonym
    recruiter::Pseudonym
    braider::Pseudonym
    proposer::Pseudonym 
    collector::Pseudonym
    timestamp::Union{DateTime, Nothing} = nothing
    signature::Union{Signature, Nothing} = nothing
end

Represents a deme configuration parameters issued by the guardian.

uuid::UUID an unique random generated community identifier;

title::String a community name with which deme is represented;

crypto::CryptoSpec cryptographic parameters for the deme;

guardian::Pseudonym an issuer for this demespec file. Has authorithy to set a roster:

recorder::Pseudonym an authorithy which has rights to add new transactions and is responsable for braidchain's ledger integrity. Issues Commit{ChainState};

recruiter::Pseudonym an authorithy which has rights to authorize new admissions to the deme. See Admission and Member;

braider::Pseudonym an authorithy which can do a legitimate braid jobs for other demes. See BraidWork;

proposer::Pseudonym an authorithy which has rights to issue a proposals for the braidchain. See Proposal;

collector::Pseudonym an authorithy which is repsonsable for collecting votes for proposals. This is also recorded in the proposal itself.

timestamp::Union{DateTime, Nothing} time when signature is being issued;

signature::Union{Signature, Nothing} a guardian issued signature.

PeaceFounder.Model.ChainState — Type

struct ChainState
    index::Int
    root::Digest
    generator::Generator
    member_count::Int
end

Represents a current chain state metadata which is sufficient for integrity checks.

Interface: index, root, generator

PeaceFounder.Model.BraidChain — Type

struct BraidChain
    members::Set{Pseudonym}
    ledger::Vector{Transaction}
    spec::DemeSpec
    generator::Generator
    tree::HistoryTree
    commit::Union{Commit{ChainState}, Nothing}
end

Represents a braidchain ledger with it's associated state. Can be instantitated with a demespec file using BraidChain(::DemeSpec) method.

Interface: push!, record!, state, length, list, select, roll, constituents, generator, commit, commit_index, ledger, leaf, root, ack_leaf, ack_root, members, commit!